# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pizza 4P's POS Cloud API",
    "description": "The public read API for the Pizza 4P's cloud. Generated from the handlers; see ADR-0019.",
    "license": {
      "name": "LicenseRef-Proprietary",
      "identifier": "LicenseRef-Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/v1/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "`GET /v1/orders?store_id=&sales_channel=&external_reference=` — the resolution path for a caller\nwhose `submit` timed out ([ADR-0061](../../../docs/adr/0061-order-relay.md)): look up what a\nreference produced rather than resubmit. Same tenant binding as `submit`; `404` while the store\nhas not yet confirmed, or for an unknown reference.",
        "operationId": "look_up_order",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "description": "The store's 26-character ULID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sales_channel",
            "in": "query",
            "description": "The channel wire token, e.g. SALES_CHANNEL_QR",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "external_reference",
            "in": "query",
            "description": "The caller's idempotency reference",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order as the store confirmed it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "A query parameter is missing or malformed"
          },
          "401": {
            "description": "The API key is missing, malformed, or invalid"
          },
          "403": {
            "description": "The API key lacks the place_orders scope"
          },
          "404": {
            "description": "No such store, or the store has not confirmed the order yet"
          },
          "503": {
            "description": "The intake is unreachable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "`POST /v1/orders` — authenticate, bind the store to the caller's tenant, and submit.",
        "operationId": "submit_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "An idempotent repeat of an already-submitted order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "201": {
            "description": "The order was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "A field is malformed"
          },
          "401": {
            "description": "The API key is missing, malformed, or invalid"
          },
          "403": {
            "description": "The API key lacks the place_orders scope"
          },
          "404": {
            "description": "No such store within the caller's tenant"
          },
          "409": {
            "description": "The store rejected the order (a precondition failed)"
          },
          "503": {
            "description": "The store did not confirm in time; retry the look-up"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/stores/{store_id}/rollups/daily": {
      "get": {
        "tags": [
          "rollups"
        ],
        "summary": "A store's per-trading-day activity rollups, oldest day first — answered from the materialised\nrollup, never a log scan.",
        "description": "Requires a valid API key with the `read_rollups` scope, and answers **only** for that key's\ntenant: the tenant comes from the verified grant, never the request, so a caller can read a\nstore's rollups only if the store is within its own tenant. A `store_id` outside the tenant is\nnot an error — it simply has no rollup and reads back as an empty list.",
        "operationId": "daily_rollups",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "description": "The store's 26-character ULID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily activity rollups, oldest day first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyRollup"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The store id is not a ULID"
          },
          "401": {
            "description": "The API key is missing, malformed, or invalid"
          },
          "403": {
            "description": "The API key lacks the read_rollups scope"
          },
          "503": {
            "description": "The rollup store is unreachable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "DailyRollup": {
        "type": "object",
        "description": "A day's activity for one store, derived from the event log.\n\nThe cloud's dashboards answer from rollups rather than by scanning the raw log\n(`docs/roadmap.md` P7). This first rollup counts events per store trading day and per event\ntype — enough for an activity dashboard, and it uses only envelope fields (`business_date`,\n`event_type`), so it needs no per-event-type decoding. Money- and order-shaped rollups, and the\nmaterialised tables that make the <10 ms dashboard query real, land with the dashboard slice.",
        "required": [
          "business_date",
          "total_events",
          "by_type"
        ],
        "properties": {
          "business_date": {
            "type": "string",
            "description": "The store's trading day, `YYYY-MM-DD` ([ADR-0022](../../../docs/adr/0022-events-partition-strategy.md))."
          },
          "by_type": {
            "type": "object",
            "description": "Per-event-type counts, ordered by type so the output is deterministic.",
            "additionalProperties": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "total_events": {
            "type": "integer",
            "format": "int64",
            "description": "Every event recorded on that day.",
            "minimum": 0
          }
        }
      },
      "MoneyDto": {
        "type": "object",
        "description": "An amount of money on the wire: `{\"currency_code\": \"VND\", \"amount_minor\": 150000}`.",
        "required": [
          "currency_code",
          "amount_minor"
        ],
        "properties": {
          "amount_minor": {
            "type": "integer",
            "format": "int64"
          },
          "currency_code": {
            "type": "string"
          }
        }
      },
      "OrderLineRequest": {
        "type": "object",
        "description": "One requested line.",
        "required": [
          "menu_item_id",
          "quantity_milli"
        ],
        "properties": {
          "menu_item_id": {
            "type": "string"
          },
          "modifier_menu_item_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "quantity_milli": {
            "type": "integer",
            "format": "int64"
          },
          "quoted_unit_price": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MoneyDto"
              }
            ]
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "description": "An order arriving through the public API.",
        "required": [
          "external_reference",
          "sales_channel",
          "store_id",
          "lines",
          "placed_at_ms"
        ],
        "properties": {
          "external_reference": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineRequest"
            }
          },
          "placed_at_ms": {
            "type": "integer",
            "format": "int64"
          },
          "sales_channel": {
            "type": "string"
          },
          "store_id": {
            "type": "string"
          },
          "table_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "description": "What became of a submitted order.",
        "required": [
          "order_id",
          "created",
          "total",
          "repriced",
          "awaiting_staff_confirmation"
        ],
        "properties": {
          "awaiting_staff_confirmation": {
            "type": "boolean"
          },
          "created": {
            "type": "boolean"
          },
          "order_id": {
            "type": "string"
          },
          "queue_number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "repriced": {
            "type": "boolean"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyDto"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "A scoped per-tenant API key, presented as `Authorization: Bearer pos_<id>_<secret>`."
      }
    }
  },
  "tags": [
    {
      "name": "rollups",
      "description": "Per-store, per-trading-day activity rollups."
    },
    {
      "name": "orders",
      "description": "Public order intake and idempotent look-up (ADR-0056)."
    }
  ]
}